+++ test/decorators/article_decorator_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class ArticleDecoratorTest < ActiveSupport::TestCase
+  def setup
+    @article = Article.new.extend ArticleDecorator
+  end
+
+  # test "the truth" do
+  #   assert true
+  # end
+end
